Cleaned up fix for getting stuck on corners in Game.java
This is not a perfect fix but rather a workaround, but the player should no longer get stuck on corners. When moving into a corner from an angle, it appears to favor moving to player to either the left or right side of the corner, depending on its orientation (NW, NE, SW, SE corners). In addition, sometimes the player would move into the wall more than the player width should allow, but the workaround forces the player back out of the wall. This can also replicated by moving into a corner at an angle.